ABTI_thread: reorder fields to reduce the struct size.
Reordered fields in ABTI_thread and used bitfields in ABTI_thread_attr
to reduce the size of struct.

+++ src/thread_attr.c
@@ -233,7 +233,7 @@ int ABTI_thread_attr_print(ABTI_thread_attr *p_attr)
     }
 
     printf("[");
-    printf("stacksize:%zu ", p_attr->stacksize);
+    printf("stacksize:%zu ", (size_t)p_attr->stacksize);
     printf("cb_func:%p ", p_attr->f_cb);
     printf("cb_arg:%p", p_attr->p_cb_arg);
     printf("]");
